src/data-search-agent/requirements.txt
adk>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
pyyaml>=6.0.1
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-exporter-gcp-trace>=1.6.0
opentelemetry-instrumentation-fastapi>=0.42b0
aiohttp>=3.9.1
httpx>=0.25.2
pyjwt>=2.8.0
